Add files via upload
This is my Calc Change Program, for which I received 100%, with the addition of 10 points extra credit, for a total of 55 out of 50 points. 

This program will ask the user for the number of quarters, dimes, nickels, and pennies s/he has and then compute the total value.  The total value of the change should be shown as the  number of dollars (i.e., a whole number) and then a separate number of cents (also as a whole number).    The assignment parameters are:

a)	Develop the money calculating program discussed below using a command line interface, where the class file includes a “main” method to launch the program, imports for required external packages, uses global and local variables where appropriate, and has additional method calls for processing.
b)	The program should use at least one method beyond main: to obtain and validate the input values for each coin type (specifically, the coin counts must be non-negative values)
c)	Include a loop in the process so that the user may enter different sets of coins (i.e., be able to run the calculation multiple times)

Extra Credit (5 points):
	Write a pseudo-code document for some task that you are familiar with but others may not be (such as cooking a particular dish, repairing an item, etc.)