Add files via upload
This is my Financials project, which is designed to calculate either an annuity or a simple interest loan. I receieved 100% on this project. A detailed description of the project is added below if you are interested.




Financials - OOP (100 points w/ 5 Extra Credit)						             2017
Two related financial processes are the calculation of an Annuity, and the calculation of a simple interest Loan.  The final balance of an annuity is based on a consistent, regular monthly deposit that builds value at a specified interest rate; the ‘logical reverse’ of this is a loan, which involves paying a monthly amount to buy down an original principal, which is also charging interest at a specified rate.  Both actions involve a monthly amount being adjusted based on an interest rate over a specified term (number of months).  

You will build a single ‘view’ form which allows input of the necessary values for either operation, and which then calls an appropriate business logic class depending on analysis choice: one class for calculating an annuity, a different class for calculating a loan payoff.  In either operation (Annuity or Loan) the user may request a ‘Full Schedule’ which will produce a second, pop-up form with the month-by-month results of the action.

Part A of the assignment involves building the main view form along with the Annuity business class (including the Annuity schedule).  Part B is to create a separate business logic class for the Loan Payment option along with the Loan Payment schedule.